requirements.txt
Django==4.1.7
djangorestframework==3.13.1
pillow==9.4.0
psycopg2-binary==2.9.5